+++ app/helpers/rails_url_shortener/urls_helper.rb
@@ -8,12 +8,13 @@ module RailsUrlShortener
     # this method return a short url or the original url if something is bad
     # Usage:
     # short_url("https://tools.ietf.org/search/rfc2616#section-5.3")
-    def short_url(url, owner: nil, key: nil, expires_at: nil, category: nil, url_options: {})
+    def short_url(url, owner: nil, key: nil, kind: nil, expires_at: nil, category: nil, url_options: {})
       # generate
       url_object = Url.generate(
         url,
         owner: owner,
         key: key,
+        kind: kind,
         expires_at: expires_at,
         category: category
       )