+++ test/dummy/db/schema.rb
@@ -10,7 +10,7 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema[7.0].define(version: 2025_05_08_120951) do
+ActiveRecord::Schema[7.2].define(version: 2026_04_16_083609) do
   create_table "rails_url_shortener_ipgeos", force: :cascade do |t|
     t.string "ip"
     t.string "country"
@@ -40,6 +40,8 @@ ActiveRecord::Schema[7.0].define(version: 2025_05_08_120951) do
     t.datetime "expires_at"
     t.datetime "created_at", null: false
     t.datetime "updated_at", null: false
+    t.string "kind"
+    t.index ["owner_type", "owner_id", "kind"], name: "index_urls_on_owner_and_kind"
     t.index ["owner_type", "owner_id"], name: "index_rails_url_shortener_urls_on_owner"
   end